requirements.txt
openai>=1.27
streamlit>=1.34
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
langchain>=0.3.0
langchain-text-splitters>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
chromadb>=0.5.0
pypdf>=4.0.2
python-dotenv>=1.0.1
tiktoken>=0.5.2
fpdf2>=2.7.0